reorganize grid widget file locations
