+++ src/grid/Row.ts
@@ -3,10 +3,10 @@ import { v, w } from '@dojo/framework/widget-core/d';
 import { DNode } from '@dojo/framework/widget-core/interfaces';
 import ThemedMixin, { theme } from '@dojo/framework/widget-core/mixins/Themed';
 
-import { ColumnConfig } from './../interfaces';
+import { ColumnConfig } from './interfaces';
 import Cell from './Cell';
-import * as fixedCss from '../styles/row.m.css';
-import * as css from '../../theme/grid-row.m.css';
+import * as fixedCss from './styles/row.m.css';
+import * as css from '../theme/grid-row.m.css';
 
 export interface RowProperties {
 	id: number;
