+++ src/grid/tests/unit/all.ts
@@ -1,4 +1,6 @@
-import './index';
-import './processes';
-import './widgets/all';
-import './utils';
+import './Body';
+import './Cell';
+import './Footer';
+import './Header';
+import './Grid';
+import './Row';
