--- src/grid/tests/unit/widgets/all.ts
@@ -1,6 +1,0 @@
-import './Body';
-import './Cell';
-import './Footer';
-import './Header';
-import './Grid';
-import './Row';
